+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- PR [#150](https://github.com/marinasundstrom/CheckedExceptions/pull/150) Fix crash in "Add catch clause" 
+
 ## [1.6.4] - 2025-07-28
 
 ### Added